Make the admin account more secure
This is done by having the check of the profile name on the profile name that is extracted from the environment variables. This cannot be spoofed by any person because the program then verifies that this current user is the correct user account (i.e. it checks whether if the user has permissions in the current profile folder (C:\User\<ProfileName>) and this is automatically locked for access by only the user whom this profile belongs to.